refactor: styled-component를 scss로 마이그레이션하라
- styled-component를 scss로 마이그레이션
- scss 코드 포맷터 적용
