+++ components/about/countryItem/index.module.scss
@@ -1,0 +1,22 @@
+@import "/styles/main.scss";
+
+.countryItemWrapper {
+  @include text("headline/500");
+
+  letter-spacing: -0.012em;
+  cursor: pointer;
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  padding: 12px 16px;
+  background: color("white");
+  border-bottom: 1px solid color("black");
+  gap: 16px;
+  transition: background-color 0.1s ease-in-out;
+
+  @media (hover: hover) and (pointer: fine) {
+    &:hover {
+      background-color: color("gray.100");
+    }
+  }
+}
