+++ components/about/countryList/index.module.scss
@@ -1,0 +1,8 @@
+@import "/styles/main.scss";
+
+.countryListWrapper {
+  user-select: none;
+  list-style: none;
+  margin: 0;
+  padding: 0;
+}
