+++ components/about/searchCountry/index.module.scss
@@ -1,0 +1,59 @@
+@import "/styles/main.scss";
+
+.searchCountryHeaderWrapper {
+  position: sticky;
+  top: 0;
+}
+
+.logoWrapper {
+  display: flex;
+  flex-direction: row;
+  justify-content: center;
+  padding: 0px 16px 24px;
+  background-color: color("gray.100");
+  width: 100%;
+}
+
+.searchInputWrapper {
+  position: relative;
+
+  &:focus-within {
+    & > svg > path {
+      fill: color("purple.500");
+    }
+  }
+}
+
+.searchIcon {
+  cursor: pointer;
+  position: absolute;
+  top: 16px;
+  left: 16px;
+}
+
+.searchInput {
+  @include text("headline/500");
+  letter-spacing: -0.012em;
+
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  height: 56px;
+  background-color: color("white");
+  width: 100%;
+  border-width: 1px 0px;
+  border-style: solid;
+  border-radius: 0;
+  border-color: color("black");
+  padding: 12px 16px 12px 56px;
+  outline: none;
+
+  &::placeholder {
+    color: color("gray.400");
+  }
+
+  &:focus {
+    caret-color: color("purple.500");
+    border-width: 2px 0px;
+  }
+}
