+++ components/common/Layout/index.module.scss
@@ -1,0 +1,54 @@
+@import "/styles/main.scss";
+
+.layoutWrapper {
+  max-width: 100%;
+  width: 100%;
+  padding: 0;
+  margin: 0;
+  height: 100%;
+  background-color: #ebedf0;
+}
+
+.contentBox {
+  height: auto;
+  width: auto;
+  max-width: 430px;
+  min-height: calc(var(--vh, 1vh) * 100);
+  position: relative;
+  background-color: color("gray.100");
+  margin: 0 auto;
+  padding: 0;
+
+  &::before {
+    z-index: 10;
+    content: " ";
+    position: absolute;
+    left: 0;
+    top: 0;
+    bottom: 0;
+    border-left: 1px solid #d2c6c6;
+  }
+
+  &::after {
+    z-index: 10;
+    content: " ";
+    position: absolute;
+    right: 0;
+    top: 0;
+    bottom: 0;
+    border-right: 1px solid #d2c6c6;
+  }
+
+  @media (max-width: 430px) {
+    width: 100%;
+    max-width: initial;
+
+    &::before {
+      content: none;
+    }
+
+    &::after {
+      content: none;
+    }
+  }
+}
