+++ styles/global.scss
@@ -1,5 +1,18 @@
-@import './variables.scss';
-@import './mixin/text.scss';
+@import "./variables.scss";
+@import "./mixin/text.scss";
+
+html {
+  -webkit-font-smoothing: antialiased;
+  -moz-osx-font-smoothing: grayscale;
+  font-family: "Pretendard Variable", Pretendard, -apple-system,
+    BlinkMacSystemFont, system-ui, Roboto, "Helvetica Neue", "Segoe UI",
+    "Apple SD Gothic Neo", "Noto Sans KR", "Malgun Gothic", "Apple Color Emoji",
+    "Segoe UI Emoji", "Segoe UI Symbol", sans-serif;
+}
+
+* {
+  box-sizing: border-box;
+}
 
 div,
 input,
@@ -24,12 +37,31 @@ button {
   }
 }
 
-html {
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-  font-family: "Pretendard Variable", Pretendard, -apple-system, BlinkMacSystemFont, system-ui, Roboto, "Helvetica Neue", "Segoe UI", "Apple SD Gothic Neo", "Noto Sans KR", "Malgun Gothic", "Apple Color Emoji", "Segoe UI Emoji", "Segoe UI Symbol", sans-serif;
+[data-rsbs-overlay],
+[data-rsbs-backdrop],
+[data-rsbs-root]:after {
+  z-index: 10;
+  max-width: 430px;
+  margin: 0 auto;
 }
 
-* {
-  box-sizing: border-box;
+[data-rsbs-backdrop] {
+  background-color: transparent;
+}
+
+[data-rsbs-overlay] {
+  border-top-left-radius: 32px;
+  border-top-right-radius: 32px;
+  box-shadow: 0px -8px 32px rgba(19, 17, 24, 0.12),
+    0px 1px 3px rgba(19, 17, 24, 0.16);
+}
+
+[data-rsbs-header] {
+  padding-bottom: 12px;
+}
+
+[data-rsbs-header]:before {
+  width: 96px;
+  height: 5px;
+  border-radius: 4px;
 }
