+++ styles/main.scss
@@ -1,0 +1,2 @@
+@import "./mixin/text.scss";
+@import "./mixin/functions.scss";
