+++ styles/mixin/text.scss
@@ -1,68 +1,56 @@
-@import '../functions.scss';
+@import "./functions.scss";
 
-@mixin text($text, $color-code: '') {
-  $list: str-split($text, '/');
-  $weight: '';
-  $font: '';
+@mixin text($text, $color-code: "") {
+  $list: str-split($text, "/");
+  $weight: "";
+  $font: "";
 
   @if length($list) == 2 {
     $font: nth($list, 1);
     $weight: nth($list, 2);
   }
 
-  @if $font == 'display' {
+  @if $font == "display" {
     font-size: var(--display-font-size);
     line-height: var(--display-line-height);
-  }
-
-  @else if $font == 'headline' {
+  } @else if $font == "headline" {
     font-size: var(--headline-font-size);
     line-height: var(--headline-line-height);
-  }
-
-  @else if $font == 'title.large' {
+  } @else if $font == "title.large" {
     font-size: var(--large-title-font-size);
     line-height: var(--large-title-line-height);
-  }
-
-  @else if $font == 'title.medium' {
+  } @else if $font == "title.medium" {
     font-size: var(--medium-title-font-size);
     line-height: var(--medium-title-line-height);
-  }
-
-  @else if $font == 'title.small' {
+  } @else if $font == "title.small" {
     font-size: var(--small-title-font-size);
     line-height: var(--small-title-line-height);
-  }
-
-  @else if $font == 'body.large' {
+  } @else if $font == "body.large" {
     font-size: var(--large-body-font-size);
     line-height: var(--large-body-line-height);
-  }
-
-  @else if $font == 'body.medium' {
+  } @else if $font == "body.medium" {
     font-size: var(--medium-body-font-size);
     line-height: var(--medium-body-line-height);
-  }
-
-  @else if $font == 'body.small' {
+  } @else if $font == "body.small" {
     font-size: var(--small-body-font-size);
     line-height: var(--small-body-line-height);
   }
 
-  @if $weight == '400' {
+  @if $weight == "400" {
     font-weight: var(--font-weight-regular);
-  }
-
-  @else if $weight == '500' {
+  } @else if $weight == "500" {
     font-weight: var(--font-weight-medium);
-  }
-
-  @else if $weight == '600' {
+  } @else if $weight == "600" {
     font-weight: var(--font-weight-semibold);
-  }
-
-  @else if $weight == '700' {
+  } @else if $weight == "700" {
     font-weight: var(--font-weight-bold);
   }
+
+  @if $color-code {
+    color: color($color-code);
+  }
+}
+
+@mixin text-color($color-code) {
+  color: color($color-code);
 }
