+++ styles/variables.scss
@@ -40,13 +40,13 @@
   /* purple */
   --purple-000: #fdfcfe;
   --purple-100: #f5f2ff;
-  --purple-200: #E1DDFE;
-  --purple-300: #C4BAFD;
-  --purple-400: #9487F8;
-  --purple-500: #6B56F0;
-  --purple-600: #523ACB;
-  --purple-700: #41319B;
-  --purple-800: #382E76;
+  --purple-200: #e1ddfe;
+  --purple-300: #c4bafd;
+  --purple-400: #9487f8;
+  --purple-500: #6b56f0;
+  --purple-600: #523acb;
+  --purple-700: #41319b;
+  --purple-800: #382e76;
   --purple-900: #211943;
 }
