+++ tsconfig.json
@@ -34,7 +34,7 @@
     "**/*.ts",
     "**/*.tsx",
     "@types",
-    ".next/types/**/*.ts"
+    ".next/types/**/*.ts",
   ],
   "exclude": [
     "node_modules",
